Reapply "feat(unshielded): the protocol version rides the indexer's progress frames, replacing the version watcher"
This reverts 618783ee, restoring b2a6969d and 1a18f24c: the indexer schema pin
and codegen at 25da0487, the compose indexer image that matches that floor, the
unshielded source selecting `protocolVersion` off `UnshieldedTransactionsProgress`,
the deletion of the polling version watcher, and the removal of the
`UnshieldedTransactionTip` document the probes read through.

618783ee backed all of this out for one reason, and it was not a defect in the
adoption: the 4.4.0-rc.2-25da0487 indexer build sits on midnight-ledger
9.1.0.0-rc.4, while this repo's wallets proved with ledger-v9 1.0.0-rc.3. The
first fee-paying transaction crashed the indexer's chain-indexer task on re-apply
("dust spend proof failed to verify ... disagreement on dust state"), which is a
ledger-line mismatch between wallet and indexer, not a progress-frame problem.

The next commit resolves that mismatch by moving the whole platform onto the rc.4
line — ledger-v9 1.0.0-rc.4 in the wallets, node 2.1.0-beta.1 under the indexer —
so the schema floor this commit declares and the ledger the indexer runs are the
same release again. The back-out's own instruction was to re-apply the adoption
by reverting it once the platform aligned on rc.4 end to end; that is this branch.
