+++ packages/indexer-client/schema.lock
@@ -3,5 +3,5 @@
 # To change where the schema is fetched from, edit schema.config.yml (repo/path), not this file.
 #   yarn schema:sync --filter=@midnightntwrk/wallet-sdk-indexer-client -- --tag <version>
 
-tag: f06af4b4
-sha256: 897f29d0cf11130470736392db3ce7b58ed3752b5660f929b0eb7e4de848de67
+tag: 25da0487
+sha256: 13a0c1606b4ec78bed266d502a143c065c04153889609b15d36afa24d5b2229f
