+++ packages/indexer-client/src/graphql/subscriptions/UnshieldedTransactions.ts
@@ -67,6 +67,7 @@ export const UnshieldedTransactions = Subscription.make(
         ... on UnshieldedTransactionsProgress {
           type: __typename
           highestTransactionId
+          protocolVersion
         }
       }
     }
