+++ packages/indexer-client/src/graphql/subscriptions/index.ts
@@ -15,7 +15,6 @@ export * from './DustLedgerEventTip.js';
 export * from './DustLedgerEvents.js';
 export * from './DustNullifierTransactions.js';
 export * from './ShieldedTransactions.js';
-export * from './UnshieldedTransactionTip.js';
 export * from './UnshieldedTransactions.js';
 export * from './ZswapEventTip.js';
 export * from './ZswapEvents.js';
